fix: change totalPrice to double in transaction

+++ src/main/java/org/kys/bnmo/model/Transaction.java
@@ -31,7 +31,7 @@ public class Transaction implements Serializable {
         this(new Customer(), new InventoryItem(), 0, 0, new Date());
     }
 
-    public Transaction(Customer customer, InventoryItem item, int quantity, int totalPrice, Date date) {
+    public Transaction(Customer customer, InventoryItem item, int quantity, double totalPrice, Date date) {
         this.transactionID = UUID.randomUUID();
         this.customer = customer;
         this.item = item;
